Safari-611.1.10/LayoutTests/fast/forms/listbox-selection-after-typeahead-expected.txt

Test checking if active selection is cleared after type ahead

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mouseDownAtOption(1);               bitPatternForSelectedOptions() is "0100000"
PASS sendkeyDown("5");                   bitPatternForSelectedOptions() is "0000010"
PASS sendkeyDown("upArrow", "shiftKey"); bitPatternForSelectedOptions() is "0000110"
PASS successfullyParsed is true

TEST COMPLETE
